Update README.md
In this project, I made a location up to 50 km. And with django backend I made an algorithm that calculates the shortest path from one point to another. With docker, I have made it shared across all operating systems. With docker compose, I made it so that several servers open at once. Nginx was used to ensure that less load came to the servers when they were opened.